backend/pyproject.toml
[project]
name = "gymio-backend"
version = "2.0.0"
description = "GYMIO smart home gym backend"
authors = [{name = "Laith Al-Saadoon", email = "alsaadoonlaith@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.4",
    "fastapi[all]>=0.115",
    "sqlalchemy>=2.0.35",
    "bleak>=0.22",
    "soco>=0.30",
    "boto3>=1.35",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
]
pi = [
    "gpiozero>=2.0.1",
    "rpi-gpio>=0.7.1",
]

[tool.hatch.build.targets.wheel]
packages = ["server", "trainer", "lights", "button", "database", "my_types", "ble_system", "sonos", "ai"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
